requirements.txt
# Librerías para el chatbot de LangChain con Gemini y Qdrant
langchain
langchain-community
langgraph
google-generativeai
qdrant-client
python-dotenv
openai
# Para manejo de API
fastapi
uvicorn
requests

# Manejo PDFs
pypdf

# Para embeddings y transformaciones
numpy
sentence-transformers

# Validación de entrada de datos
pydantic
